+++ CHANGELOG.md
@@ -5,7 +5,13 @@
 
 <br>
 
-## v2.0.0 (date here)
+## v2.0.0 (8/4/2021)
 - Added the `list` command
 - Added the `config` command
 - Added support for gitignores and licenses
+
+<br>
+
+## v2.0.1 (15/4/2021)
+- Made a separate 'list.go' file for the `list` command
+- CI build bug fix (redefined embed paths for licenses and gitignores)